Tweak the RocksDB C++ build to recover the ability to compile in parallel
The latest version 6.15.4 is way WAY slower to compile than the previous
version, so that the hack we use that #include's all source files from
one big C++ file (and therefore runs the entire compile on a single CPU
core) was becoming a problem.

I found another way to do this that preserves the workaround behavior we
need but also results in multiple source files to be compiled, so the
build time is much better now.